+++ packages/drizzle/CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## next (YYYY-MM-DD)
 
+* Refactor: remove `DrizzleResolverFactory.create`, use `drizzleResolverFactory` instead
+
 ## 0.7.2 (2025-02-16)
 
 * Fix: update document link in README
