+++ packages/drizzle/src/index.ts
@@ -198,5 +198,4 @@ export type TableSilk<TTable extends Table> = TTable &
     >
   }
 
-export * from "./input-factory"
-export * from "./resolver-factory"
+export * from "./factory"
