+++ packages/drizzle/test/resolver-postgres.spec.ts
@@ -9,7 +9,7 @@ import {
 import { type YogaServerInstance, createYoga } from "graphql-yoga"
 import { afterAll, beforeAll, describe, expect, it } from "vitest"
 import { config } from "../env.config"
-import { DrizzleResolverFactory } from "../src"
+import { drizzleResolverFactory } from "../src"
 import { post, postsRelations, user, usersRelations } from "./schema/postgres"
 
 const schema = {
@@ -48,8 +48,8 @@ describe("resolver by postgres", () => {
   beforeAll(async () => {
     try {
       db = drizzle(config.postgresUrl, { schema })
-      const userFactory = DrizzleResolverFactory.create(db, "drizzle_user")
-      const postFactory = DrizzleResolverFactory.create(db, "drizzle_post")
+      const userFactory = drizzleResolverFactory(db, "drizzle_user")
+      const postFactory = drizzleResolverFactory(db, "drizzle_post")
       gqlSchema = weave(
         userFactory.resolver({ name: "user" }),
         postFactory.resolver({ name: "post" })
